pyproject.toml
[tool.poetry]
name = "code-builders"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
psycopg2 = "^2.8.6"
razorpay = "^1.2.0"
ipython = "^7.18.1"
django-phonenumber-field = "^5.0.0"
phonenumbers = "^8.12.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
